Move Iteraction->Context logic to handle_resp.py
